Add some string operator functions
Add some string operator functions that can convert raw input text to rich text format text of a polynomial. The test code are provided too.  And the string operators are dependent on Boost::tokenizer. The next is to complete the string operator functions and the solution for  getting unit step response.